refactor(transforms): break down large transform files into focused modules
Split catalyst-prefix.ts (669 lines) and semantic-colors.ts (318 lines) into smaller, focused modules for better maintainability and debugging:

• catalyst-prefix.ts → 4 modules:
  - catalyst-prefix-core.ts: AST setup and mapping system
  - catalyst-prefix-imports.ts: Import declaration processing
  - catalyst-prefix-references.ts: Reference transformations
  - catalyst-prefix-types.ts: Type alias declarations

• semantic-colors.ts → 2 modules:
  - semantic-colors-mappings.ts: Component detection and color patterns
  - semantic-colors-transform.ts: Core transformation logic

Preserves all existing documentation and phase comments. All 527 tests passing.
